Divide code into smaller modules
