+++ app/parse.py
@@ -1,11 +1,8 @@
 import json
 import logging
-import os
 import re
-import time
-from dataclasses import dataclass, asdict
-from functools import wraps
-from typing import Callable, Any, List
+from dataclasses import dataclass
+from typing import Any, List
 from urllib.parse import urljoin
 
 import pandas as pd
@@ -20,10 +17,8 @@ from openpyxl import load_workbook
 from openpyxl.styles import Alignment
 from openpyxl.worksheet.hyperlink import Hyperlink
 
-BASE_URL = "https://mate.academy/"
-APP_DIR = os.path.dirname(os.path.abspath(__file__))
-JSON_RESULT_FILE = os.path.join(APP_DIR, "courses_data.json")
-EXCEL_RESULT_FILE = os.path.join(APP_DIR, "courses_data.xlsx")
+from config import BASE_URL
+from logger import log_time
 
 
 class HTTPResponseError(Exception):
@@ -50,40 +45,6 @@ class CourseModule:
 @dataclass(frozen=True)
 class CourseDetail:
     modules: List[CourseModule]
-
-
-def configure_logging() -> None:
-    logger = logging.getLogger()
-    logger.setLevel(logging.DEBUG)
-
-    file_handler = logging.FileHandler("app.log")
-    file_handler.setLevel(logging.INFO)
-    file_handler.setFormatter(
-        logging.Formatter("%(asctime)s - %(levelname)s - %(message)s")
-    )
-    logger.addHandler(file_handler)
-
-    console_handler = logging.StreamHandler()
-    console_handler.setLevel(logging.DEBUG)
-    console_handler.setFormatter(
-        logging.Formatter("%(asctime)s - %(levelname)s - %(message)s")
-    )
-    logger.addHandler(console_handler)
-
-
-def log_time(func: Callable) -> Callable:
-    @wraps(func)
-    def wrapper(url: str, *args: Any, **kwargs: Any) -> Any:
-        start_time = time.time()
-        result = func(url, *args, **kwargs)
-        elapsed_time = time.time() - start_time
-        logging.info(
-            f"Time taken by {func.__name__} "
-            f"for {url}: {elapsed_time:.2f} seconds"
-        )
-        return result
-
-    return wrapper
 
 
 @log_time
@@ -428,42 +389,3 @@ def adjust_column_width(worksheet: Any, col: Any) -> None:
             pass
     adjusted_width = min((max_length + 2), 50)
     worksheet.column_dimensions[column].width = adjusted_width
-
-
-@log_time
-def main(base_url: str) -> None:
-    configure_logging()
-    try:
-        courses = get_all_courses(base_url)
-        courses_data = []
-        for course in courses:
-            course_dict = {
-                "name": course.name,
-                "link": course.link,
-                "description": course.description,
-            }
-            (
-                details,
-                num_modules,
-                num_topics,
-                full_time_duration,
-                flex_time_duration
-            ) = get_course_detail(course.link)
-            course_dict["details"] = asdict(details)
-            course_dict["num_modules"] = num_modules
-            course_dict["num_topics"] = num_topics
-            course_dict["full_time_duration"] = full_time_duration
-            course_dict["flex_time_duration"] = flex_time_duration
-            courses_data.append(course_dict)
-    except HTTPResponseError as e:
-        logging.error(f"Failed to fetch courses: {e}")
-    else:
-        write_to_json(courses_data, JSON_RESULT_FILE)
-        write_to_excel(courses_data, EXCEL_RESULT_FILE)
-        logging.info(
-            f"Courses data saved to {JSON_RESULT_FILE} and {EXCEL_RESULT_FILE}"
-        )
-
-
-if __name__ == "__main__":
-    main(BASE_URL)
